feat: moving color picker into its own component, adding colors array
