+++ client/src/assets/colorPickerColors.js
@@ -1,0 +1,30 @@
+export const colors = [
+  // red and pink
+  '#993955',
+  '#B02E0C',
+  '#F4989C',
+
+  // orange
+  '#EB4511',
+  '#FA9F42',
+
+  // yellow
+  '#FFD166',
+  '#FFED65',
+  '#F6E7CB',
+
+  // green
+  '#8CB369',
+  '#A6D3A0',
+  '#B0DB43',
+
+  // blue
+  '#4A6FA5',
+  '#91ADC2',
+  '#ACECF7',
+
+  // purple
+  '#89608E',
+  '#9B7EDE',
+  '#DAC4F7',
+]
